requirements.txt
streamlit>=1.24.0
crewai>=0.1.0
langchain>=0.0.300
openai>=0.27.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0
pytest>=7.0.0
pysqlite3-binary>=0.5.2
chromadb>=0.4.18